Refactor backend.text's package structure (#8492)
* Rename backend.text.admin package to backend.text.metadata

* Rename backend.text.database package to backend.text.data

* Add schema package for backend handler

* Refactor DatabaseBackendHandler

* Refactor UseDatabaseBackendHandler

* Refactor UseDatabaseBackendHandler
