+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/test/java/org/apache/shardingsphere/proxy/backend/text/metadata/ShowDatabasesBackendHandlerTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.backend.text.admin;
+package org.apache.shardingsphere.proxy.backend.text.metadata;
 
 import org.apache.shardingsphere.infra.auth.Authentication;
 import org.apache.shardingsphere.infra.auth.ProxyUser;
@@ -27,8 +27,11 @@ import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
 import org.apache.shardingsphere.proxy.backend.communication.jdbc.connection.BackendConnection;
 import org.apache.shardingsphere.proxy.backend.context.ProxyContext;
 import org.apache.shardingsphere.proxy.backend.response.header.query.QueryResponseHeader;
+import org.apache.shardingsphere.proxy.backend.text.metadata.schema.ShowDatabasesBackendHandler;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import java.lang.reflect.Field;
 import java.sql.SQLException;
@@ -40,35 +43,31 @@ import java.util.Properties;
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
-import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public final class ShowTablesBackendHandlerTest {
+@RunWith(MockitoJUnitRunner.class)
+public final class ShowDatabasesBackendHandlerTest {
     
     private static final String SCHEMA_PATTERN = "schema_%s";
     
-    private ShowTablesBackendHandler tablesBackendHandler;
+    private ShowDatabasesBackendHandler showDatabasesBackendHandler;
     
     @Before
-    public void setUp() throws NoSuchFieldException, IllegalAccessException {
+    public void setUp() throws IllegalAccessException, NoSuchFieldException {
         BackendConnection backendConnection = mock(BackendConnection.class);
         when(backendConnection.getUsername()).thenReturn("root");
-        tablesBackendHandler = new ShowTablesBackendHandler(backendConnection);
-        Map<String, ShardingSphereMetaData> metaDataMap = getMetaDataMap();
-        when(backendConnection.getSchemaName()).thenReturn(String.format(SCHEMA_PATTERN, 0));
+        showDatabasesBackendHandler = new ShowDatabasesBackendHandler(backendConnection);
         Field metaDataContexts = ProxyContext.getInstance().getClass().getDeclaredField("metaDataContexts");
         metaDataContexts.setAccessible(true);
-        metaDataContexts.set(ProxyContext.getInstance(), 
-                new StandardMetaDataContexts(metaDataMap, mock(ExecutorEngine.class), getAuthentication(), new ConfigurationProperties(new Properties()), new MySQLDatabaseType()));
+        metaDataContexts.set(ProxyContext.getInstance(), new StandardMetaDataContexts(
+                getMetaDataMap(), mock(ExecutorEngine.class), getAuthentication(), new ConfigurationProperties(new Properties()), new MySQLDatabaseType()));
     }
     
     private Map<String, ShardingSphereMetaData> getMetaDataMap() {
         Map<String, ShardingSphereMetaData> result = new HashMap<>(10);
         for (int i = 0; i < 10; i++) {
-            ShardingSphereMetaData metaData = mock(ShardingSphereMetaData.class, RETURNS_DEEP_STUBS);
-            when(metaData.isComplete()).thenReturn(false);
-            result.put(String.format(SCHEMA_PATTERN, i), metaData);
+            result.put(String.format(SCHEMA_PATTERN, i), mock(ShardingSphereMetaData.class));
         }
         return result;
     }
@@ -81,17 +80,17 @@ public final class ShowTablesBackendHandlerTest {
     }
     
     @Test
-    public void assertExecuteShowTablesBackendHandler() throws SQLException {
-        QueryResponseHeader actual = (QueryResponseHeader) tablesBackendHandler.execute();
+    public void assertExecuteShowDatabaseBackendHandler() {
+        QueryResponseHeader actual = (QueryResponseHeader) showDatabasesBackendHandler.execute();
         assertThat(actual, instanceOf(QueryResponseHeader.class));
         assertThat(actual.getQueryHeaders().size(), is(1));
     }
     
     @Test
-    public void assertShowTablesUsingStream() throws SQLException {
-        tablesBackendHandler.execute();
-        while (tablesBackendHandler.next()) {
-            assertThat(tablesBackendHandler.getRowData().size(), is(1));
+    public void assertShowDatabaseUsingStream() throws SQLException {
+        showDatabasesBackendHandler.execute();
+        while (showDatabasesBackendHandler.next()) {
+            assertThat(showDatabasesBackendHandler.getRowData().size(), is(1));
         }
     }
 }
